+++ MoltenVKShaderConverter/include/MoltenVKSPIRVToMSLConverter/SPIRVConversion.h
@@ -1,1 +1,0 @@
-../../MoltenVKSPIRVToMSLConverter/SPIRVConversion.h